cmake: signed strlen-zero idiom in CmakeNameCtrl (89.1%)

+++ src/cmake.cpp
@@ -2764,8 +2764,8 @@ int CMenuPcs::CmakeNameCtrl()
             } else if ((down & 0x100) != 0) {
                 short curRow = CmakeState(this)->m_row;
                 if (curRow >= 5) {
-                    unsigned int emptyLen = strlen(s_CmakeInfo.m_name);
-                    if ((emptyLen & (static_cast<unsigned int>(-emptyLen | emptyLen) >> 31)) == 0) {
+                    int emptyLen = strlen(s_CmakeInfo.m_name);
+                    if ((emptyLen & ((-emptyLen | emptyLen) >> 31)) == 0) {
                         Sound.PlaySe(4, 0x40, 0x7F, 0);
                         return 0;
                     }
